requirements.txt
aac-datasets==0.7.0
accelerate==1.10.1
datasets==4.1.1
diffusers==0.35.1
easydict==1.13
einops==0.8.1
ftfy
hf-xet==1.1.9
huggingface-hub==0.34.4
imageio
ipykernel
joblib==1.5.2
jupyterlab
jupyterlab_widget
laion_clap==1.1.7
lazy_loader==0.4
matplotlib==3.10.6
ninja
numba==0.61.2
numpy==2.2.6
omegaconf
openai-clip
opencv-python==4.12.0.88
pandas==2.2.3
pillow==11.0.0
protobuf==6.32.1
pyarrow==21.0.0
pycocoevalcap
pycocotools
pytorch-fid==0.3.0
regex
resampy==0.4.3
safetensors==0.6.2
scipy
sentence-transformers==5.1.0
sentencepiece==0.2.0
sentry-sdk==2.38.0
six==1.17.0
timm==1.0.19
tokenizers==0.22.1
torch==2.8.0+cu126
torchaudio==2.8.0+cu126
torchdiffeq==0.2.5
torchlibrosa==0.1.0
torchvision==0.23.0+cu126
tqdm
transformers==4.56.1
triton==3.4.0
wandb==0.21.4
wav2clip==0.1.0
xformers==0.0.32.post2
